Python/django_rest_framework_ejemplo1/requirements.txt
asgiref==3.8.1
cffi==1.17.0
Django==5.2
djangorestframework==3.16.0
greenlet==0.4.13
hpy==0.9.0
readline==6.2.4.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
